Handle user provisioning better
Now, instead of a one-shot job that only fires on `monit start` (and
sometimes not even reliably then), jumpbox has a watcher script that
spins on a 5m loop checking to see if it needs to "fix" things.

This commit also introduces account invalidation and removal.  To
invalidate an account, remove it from the manifest.  To delete it, add
it to the jumpbox.delete manifest property (list).

Fixes #11